Moving from pysynphot to synphot/stsynphot so as to be compatible with pandeia 1.5.2 and be on a supported module

+++ stips/astro_image/astro_image.py
@@ -168,7 +168,7 @@ class AstroImage(object):
 
     
     def __del__(self):
-        if os.path.exists(self.fname):
+        if hasattr(self, 'fname') and os.path.exists(self.fname):
             os.remove(self.fname)
 
     def copy(self):
